backend/requirements.txt
flask
flask-cors
numpy
torch
torchvision
Pillow
